Updated CHANGELONG to include the switch connection support.
Modified the table structure for PatchConnection slightly.  There is now a notes field for both the front and rear.  Updated both the create.sql and the upgrade sql.

Added basic functions to make panel connections and to list existing ones.  Still need to add more for clean-up when you delete a device.

Addresses #73