Have unit tested the connection making and removing processes for both switches and patch panels.  Still need to write the functions for removing all connections when you delete a device that is a patch panel.
Unlike with switch connections - when you remove a patch panel connection, it simply 0's out the Device ID and Port number, and then blanks out the Notes.  This is because you have a front and rear for each port number, and just because you delete a front connection, it doesn't mean that you want to delete the rear.

Addresses #73