Added in logic in the MoveToStorage and UpdateDevice functions of the Device class to account for Stupid User Tricks.  Specifically changing the DeviceType from Switch, Patch Panel, or Chassis to anything else - because if you do so, you create orphans.
In the case of a chassis, if you have child devices assigned, the update quietly fails.  If there are no child devices, the update succeeds.

In the case of patch panels and switches, any modification of the DeviceType will cause all associated records of connections to be pruned, and then the update will go through.

Fixes #110
Fixes #111