Added a new report for emailing the Facility Manager with the installations listed as happening in the past 7 days.  The number of days is configurable (Wilbur added the config screen) in the NewInstallsPeriod variable.
Also added a variable to the config called InstallURL, since scripts called from email can't poll the current URL to determine what to put in anchors to the devices and cabinets.

report-em_new_installs.php is to be called from crontab, or other periodic scheduler.  The suggested period is weekly, on a day/time of the Facility Manager's choosing.

Also added a MakeSafe() method for the Device class.  It still needs to be populated throughout the rest of the class.