Added the ColVis extra to the DataTables, which will allow users to hide/show columns on demand.  Updated the search_export.php file to include this new extra.
Still need to add in all imaginable columns for this report so that users can deselect on demand.  This won't be a thorough export, since some things simply can't be represented in a single tabular format (such as child devices, VMs, power connections and network connections).  However, will add pretty much everything about a device that is in the main table.