Added the field for User Right of AdminOwnDevices.
Updated all of the User class methods to use PDO Library per #164

Updated the dc_stats, cabnavigator, and devices pages to base viewership on what departments a user is a member of, rather than the global ReadAccess right.

Use cabnavigator.php as the reference for how to handle the new right scheme.

#185