Expanded the table when viewing a switch to include some of the details (but still not all, yet) of the DevicePorts table.
Also added in functionality to the DevicePorts - if the device is a Switch and the DevicePorts query is empty for it, create new rows based upon the number of ports defined for that device.  It will also query the switch via SNMP if those values are defined, and pre-populate some of the values.

Also changed the SNMP calls to break out of the loop if a timeout occurs, because if you have a misconfigured agent on a 48 port switch, you will hang up your browser session for a LONG time while it tries 48 times for the names, 48 times for the alias, and 48 times for the status.

#191