Cleared the error messages from the log - wasn't really related to this issue, but since the conversation started here, referencing it.
The errors came from the ColorCoding::getCode() method.  A test was being done to see if a row could be fetched from the SQL execution and if not, an error was logged.  However, when the SQL is valid but the result set is zero, there really shouldn't be an error condition.  Removed that logging, because every port with no attached device was throwing the error.

Fixes #181