Handling of "FrontEdge" property on cabinets #375
Changes:
-Added "Air Flow" option in the Select of "dc_stats" page  for
consistency with the other informations
-Arrows of "Air Flow" are drawn with a gradient (blue = cold, red =
hot). This allows to visually distinguish hot and cold aisles
-On the "dc_stats" page you can change the "FrontEdge" of cabinets
quickly:
1. With "Air Flow" option selected...
2. with your mouse in a cabinet, if you press the "Shift" key and then
you move the mouse out of the cabinet, this displacement indicates the
airflow direction. Cabinet will be updated and the page will be
reloaded.
3. If, in addition to the "Shift", key "Ctrl" is also pressed, when you
move the mouse out of the cabinet, all the cabinets of that row are
updated. (Normally all cabinets in a row have the same orientation)
-On the "row_view" has been modified by the cabinets. To date the order
was alphabetical (or reverse). This order is also stored in the database
in table fac_CabRow. This is no longer necessary because the "FrontEdge"
field and the coordinates of each cabinet, and you can build the line as
it is in reality (regardless of how the cabinets are called). In
addition you can have a view of the row ahead and behind.