Menu Display / Navigation #415
This is a first attempt.
Works well with all pages, excluding handling large images as dc_stats.
On these pages we should limit the size scaling the image (preferably)
or by enclosing them in a div with scroll bars.
The goal would be that the divs did not exceed the size of our screen,
so no scroll bars appear on the browser.