Check the output status of the dot commands
We now check the output status of the dot commands, and provide error
messages if something isn't right. samilliken/openDCIM#486

Graphviz's bitmap renderers are easy to overflow with a large enough
graph. I believe that telling the user to use the svg output option is
reasonable if such an issue happens.