Added new assets.
Fixed horrible bug in the display of patch panels with rear connections.
The AJAX method getport would return an ABS value for portNumber.
This screwed rear connections and made edits inconsistent as the AJAX disagreed with the PHP
Now returns integer, and AJAX code modified to append (rear) to ports that connect to the rear of the connectedDevice.