Added back in usermgr.php - had meant to delete contacts.php, which this commit does.
Made the usermgr.php actually work correctly, now.  Added in a box at the bottom to let you assign departments for a user, similar to the way we assign users to departments.

Have to go off-site, so checkpointing what I have currently.  The selector box for the departments doesn't fit.

#424