Found where the port information for the last connection (first member outside of the scope) was showing as 'outside filter' instead of the real name and fixed that.
Also found that the map report was treating ports that were not connected as blank entries.  This is the case if you had ever set the port to a value and then deleted it - the ConnectedDeviceID is then set to 0, rather than null, and the report was only checking for null.  It now checks for either condition.

Fixed #519