Added in two new config options to set the global default behavior with repository interactions.
ShareToRepo - Whether or not new templates will, by default, be marked for sharing with the global repository (which will be a Site Admin initiated process).  Default is not to share.

KeepLocal - When synchronizing with the Repo, whether or not to keep local db values over any changes that may be within the repository.  Default is to keep local.

#537