New free fonts for FPDF and important bug fix in UA translation
1. New free fonts for FPDF reports and new  example loadfonts.php (it is
more preferable to use condensed fonts).
2. Translate "Device Images" in devices.php.
3. Important comment in power_pdu.php (line 411) and appropriate
correction in UA localization:
-- if translation into other language of the message "__("Are you sure
that you..." contains the character " ' ", in the po-file of
localization it is necessary to apply double escaping to it, i.e. \\' .
Otherwise, there will be impossible an execution of a call "Test
Communications" and links in the table power connections of PDU!
4. New template file crontab.tmpl with the correct syntax of execution
of all necessary php-scripts of openDCIM.