Removed the link to and the file for Sensor Templates (now part of Device Templates).
Continued the propagation of SNMPVersion code changes so that it's no longer part of the CDU and Sensor templates, since it's now a global template field.

Updated the installer to work with the new location for the SNMPVersion during the template conversion.

Removed the v2AuthName field as it's the same as SNMPCommunity.

#584
#555