Checkpointing changes - not in a working state, yet.
Since Cabinet:Sensor is no longer a 1:1 ratio, need to change the way that we handle sensor readings.  Am creating a new table to track by the device rather than the cabinet.  Will have to look at all sensors for a cabinet to get the full view.  Will likely need to add gauges on the cabnavigator page to show the average temperature and humidity among the sensors (a set for front and a set for rear).