I think I found all of the reverted changes from the GitHub screwup of yesterday.
The install process now works properly for a fresh database.  On the Complete page, the installer now has a button to go ahead and slurp down all of the Manufacturer names from the repository.  It also instructs them to subscribe to the ones that they want and then set up repository_sync.php to run either manually or as a cron job.

#555