Adding in new rack drawing functions
Still to do is correct the row view so that it will render correctly.  It isn't tagging the racks for front/rear properly yet and in turn causing the images to render incorrectly.  The side view still needs to be fleshed out with the JS to mark the areas on the cabinet as used. Finally, still need to work out a way to build the legend like the old routines.  #526