Mostly complete user side repository sync
Adding more more methods to the api. removed the manufacturer sync and all associated code from device_manufacturers. adding in a new page just for interacting with the repository that is not server bound for folks that cannot use the command line based version for whatever reason.  Still need to add in the put method for manufacturer creation to the v1 api and correct the remaining javascript problems. #537