More work on client side repo sync
Added in a change to the db schema to restrict userid to be a unique field.  Need to address this with the update as well.  Adding in methods to create power ports via the api and renamed all of the paths to be singular to be in line with the rest of the paths.  Also added in an additional click target and pop up to show the power port names to match the functionality on the data ports #537