Correct notes search problem
Following up on the email that was sent out about the notes search not functioning correctly.  I partially addressed this a while back when I adjusted the search and loose search functions extend function.  This patch is re-adding in the port notes searching ability.  Added indexes to the notes fields on the ports so we don't waste cpu time scanning through a gigantic table for no good reason.  Also added search and loose search functions to both deviceports and powerports classes to match the others