Fixing cabinet editor
Refactoring the javascript for the cabinet editor to use the API.  Also extended the API to support get methods for zones and cabinet rows.  Added the nearly standard search function to both of those classes for use by the API as well.  When zones and rows were originally added they were meant to only have rows nested in a zone but when we split that out we never fixed the editor to fully support it.  Now it should.  I hope.  Maybe. #698