More repo sync polishing
Adding in the two missing routes in the API for creating slots.  Fixed two typos from where I renamed dataports and powerports to be singular to try and be consistent.  Added in a link to the pull from repo function to mimic the local templates that are linked to make it easy on the user to see the local version quickly. #537