Layout bug on device_template screen
Not sure what caused this all of the sudden but the checkboxes for the extra parameters were on two lines, the dropdown for the selector boxes was on a second line as well.  Something added a few pixels in the layout and it was the same in both chrome and firefox.  I made a nasty hack to move the sizes around post load of the page and it seems to work.  I should probably look at this later more in depth