Adding bulk power importer
The bulk power importer uses most of the same code from the bulk network
importer except that it makes reference to the PowerPort table and the
power port class.

A new sub menu was added in the "Bulk Importer" section called "Import
Power Connections". This allows you to get to the bulk power importer
that was added.