Move export connections to dedicated function
To make the connections export more useable, I moved that
functionality to a new file - connections_spreadsheet.php. This
function also supports hiding empty power connections (off by
default.) During this change, I also fixed a bug which would result
in empty connections in the spreadsheet (presented as '0's.) That
bug was probably the result of PHPExcel's writer precalculating
cell values. I fixed it by assigning the actual strings to the
cell, not a formula which resolves to a string.

- Fix broken connections in workorder email