Add Project outage report
Added a new report 'Project Power Outage Report' which does the
same thing as the 'Simulated Power Outage Report' but for projects.
Instead of producing a table, this generates a spreadsheet. The
Summary sheet contains a list of all the projects and their
statuses. Each project also has a dedicated sheet to show each
device's status within the project. If tags are provided, the
simulation will only be run for the tagged set. If 'Only show
down/unknown devices' is checked, projects with 100% normal status
will be omitted from the summary and will not have a dedicated
sheet.