User requested changes/fixes
Added a section to the mpdf stuff to generate a known writable temp directory for mpdf to use to avoid an error.  Added in a new environment variable for the language settings to hopefully correct some new ewwwwbuntu problems.  Adjusting the height of the device labels to account for text decoration items both above and below the text.  close #814