Clearing out debug line
User was reporting a weird sql string as an error in his error logs.  I found the culprit that was a debug statement that had been left to log the sql query to verify it was working.  While I was in there I spotted two other errors and corrected them.