Update to rowview post @f39b6fa
User filed a bug report on the removing of all the php_self nonsense that was actually from the position removal attempt I made.  It was setting the cabinet widths too wide and the controls weren't uniform on the row view like they were in cabnavigator.  Made a quick check in the js to handle the differences