Lots of things wrapped up in one, here.
Fixed #843 by setting up the autoloader for Composer packages to pull in what we need for openDCIM.

Started on #842 by deleting the mpdf, PHPExcel, Swiftmailer, and Slim libraries and then adding in the composer packages.

So far I have gotten the authentication middleware working, and all of the GET routes.  The POST and PUT routes are commented out so that they don't break the app.

Have not touched the other packages, yet.