Finishing merging of custom attributes into device model
Finally got around to mostly finishing the merging of the custom attributes into the device model.  Still need to figure out all the data sanitizing that needs to be done to get the label column set to unique.  Also added in data validation to the configuration page to make sure that only letters, numbers, and underscores are used in the custom attributes so that they can be form submitted without any weirdness. #80