Adding a fallback for the first port detection
Dolemite has wanted to put in a section in the config for a while to let users alter the regex for the first port detection but I think that is going to cause more problems then the few it solves. We might still add a function to disable it outright temporarily or something but for now if we dont' get any ports back this should kick back the full list of ports and an error message to the end user to report the device to us. close #840