Update datacenter Text Output To data center In inventory.php
In user visible output the phrase should be "data center" not
"datacenter". Updating the few places the output is inconsistent is
much simpler than renaming the project openDIM.

This changeset alters the errornious comment in the css/inventory.php
file (which a user could see with view source I think).