Update datacenter Text Output To data center In en_US po File
In user visible output the phrase should be "data center" not
"datacenter". Updating the few places the output is inconsistent is
much simpler than renaming the project openDIM.

This changeset alters the en_US openDCIM.po file.  This is the file
which is used to generate the majority (all?) of the user visible text
on the web interface.