More configuration page for device statuses #945
Adding works when clicking the plus but not by hitting enter. It does not add the new entry to the list, yet. Removing works 100%. Updating a name/color works 100%.  Had to correct the class because they just didn't work and I wasn't able to debug the sql that the stupid prepared methods were fucking up somehow so I changed them to shit that works.  Added in the put/post/delete methods in the api, updated the get methods to match the put/post/delete.