Normalize models
Correcting the problem in #960 where the notes from the templates weren't getting applied to the notes on the ports upon device creation.  Also dropping the field PortNotes since we stopped using it several releases ago when we ditched the port notes in favor of a single note field that applies to the connection and port.