Added a call to update the NavTree whenever a cabinet, row, zone, datacenter, or container is created, updated, or deleted.
Renamed the existing buildNavTreeHTML function to updateNavTreeHTML that now saves to fac_DataCache instead of spitting to output buffer.

Created a new buildNavTreeHTML that simply pulls from cache.   This was to keep from messing up any calls to it that I didn't find.

Obviously, had to create a new table for data caching.

#1002