Adding in first test for optimization #1004
This is not complete code!  Be sure to add in the new table and then run create_cache_for_1004.php in a browser.  This might take a while depending on the number of objects in the database.  For production we\'ll generate this during the install/update and then it will get the individual objects updated whenever a device gets updated.  This is still PoC and just need to get values in place to see where we sit with load times.  In my dev instance a row with medium density went from 10s+ load to anwhere between 3.5 - 5s