Fix the cabinet image controls #1004
Moving the function that inserts the cabinet image controls and reads the initial state of the cookie for the rack layout to after the devices have been added to the dom.  Had to add it to both of the functions in common.js because there is no common place between them that guarantees the devices are loaded as well as the images.  Having it load from the parent page was clumsy and prone to error with the devices not being ready in the dom yet.  There is still a bug when the position is disabled and the images are disabled.  I think this has been there for a while and I argued about it when I implemented the position toggle in the first place