Adding in the device image cache hooks
Adding in the hooks into the create and update device functions to rebuild the image cache.  Also updated the WhosYourDaddy function to have the option to return not only the immediate parent but the root of the lineage.  I couldn't find anywhere in the code base we are actually using that function also.  This should make this ready for production minus putting in a cache generate function into the installer for the update. Definitely ready for someone with a large install base to give it a good testing. #1004