Adding in some css based loading placeholders
To add a little polish to the UI and let the user know that something is still going on even when the cabinet loading doesn't appear to be doing anything I've added in some loading animations that we can use for anything.  They are global javascript variables set in the common.js file.  Or we can just use the html in something if so desired.  #1004