Initial commit for changing out the import/export functions
This is nearly complete now.  The previous version did handle custom attributes which this does not at the moment, I plan to address that tomorrow.  For now you can export a template as a json download complete with pictures and import it to another system.  I need to rework the DeviceTemplate functions to mimic what we did with the device model so that the custom attributes are all part of the model itself and not some extra entity hanging out in limbo. #1007