custom attribute checkboxes were misleading
I found that in the device_templates.php we were allowing the user to try to set an attribute as required when it wasn't enabled but we are silently correcting that on the back end.  To keep the UI in sync with how this is processed I have added a few javascript hooks so that when the user unchecks enabled it will clear the checkmark, if it exists, from the required field.  Likewise if the user checks required it will automatically put a check into enabled