Changed the report a bit to simplify it.
Rather than joining the Template information in the SQL statement, just
load up an array with the entire Template table in memory.   That keeps
us from having to do a double LEFT JOIN in the SQL.   Even a large
installation should have less than 1000 templates, and the data isn't that
big in memory.

Verified with data set that was missing in prior report.

Fixed #1042