Made a quick utility to solve the problem that CenturyLink had, and adding to GitHub just in case someone else encounters the same problem.
At some point, which they think was during the upgrade to 18.01, it set all of their TemplateID=0 devices to a Height=3.   The fac_GenericLog doesn't show any user activity doing this, but it does have a different value in the logs than the current value of the DB.

This script, if run from the command line, will compare the current value in the db for all TemplateID=0 devices (other than Status="Disposed") against the last logged value, and if they mismatch, revert to the logged value.