Function redeclaration problems
Somehow we were looping when removing objects and redeclaring functions that were only supposed to be run once.  Instead of tracking down the loop which I suspect is due to the menu caching I just added some checks to not redeclare them.  The problem was evident on containers, racks, datacenters, etc, anything dealing with the navtree.   close #1047