Updating the js that sets the rack color coding
This isn't the right issue for this but I am going to close it out along with this all the same.  This is updating the InsertDevice function to correctly color devices that have something in the front that has a different owner from the rear because this was a problem for some people. I had to use the jquery :last selector because the :last-child wasn't able to target the last .pos element correctly.  There is a slight performance decrease here but it shouldn't be noticable to the end user. close #1014