Nesting dolls correction
Reworked the loop that processed devices for the search_export to have a function that can be called in a way that children of children will also get processed not just the first born.  I also reworked how we handle the custom attributes a hair so that children will show their custom values as well as parent devices.  close #996