Update redirect function to allow for proxied setups
This is going to fix the issue of people having opendcim installed on non-standard ports and then running it through a reverse proxy on another port.  There was also an issue with how this function was working with detection of proper links so this is also a bugfix.  close #1078  fixes #965  fixes #963