Update to rightscheck for configurable paths
Adding in a secondary loop that will check for paths that will be supplied by the end user addition to scott's initial commit for #976.  If we keep this we will need to remove drawings and whatever else we add as configuration items from the default list of directories being checked.