Updating image paths
Updating more locations where we had the drawings and pictures paths hard coded.  Also added in some data validation to the configuration screen to enforce storing the path with a trailing slash.  Should probably go back over all the places I just updated and make sure that we have the proper slash or no slash in the code so we don't get broken images.  #976