Bugfix
Found a place where $config was referenced in a javascript function and needed to be parsed by php for the correct value #976.  This js is going to be problematic and we are going to need to look for these.  Since we are allowing arbitrary input for the directory name it could potentially hold a single tick which would break this .html() function so these will all need to be found and converted to double quotes and escape out the ones in the html