More work on custom paths
I think I got it working now to support weird characters in the directory names as well quotes.  Obviously anything really strange is still going to break but this should work for single ticks now.  I have made one pass over the places where we pull in the picture or drawing path looking for issues with single quotes and the javascript but this needs way way more testing because that crap is all over the place.  #976