Correction to rowview oddities
Reported on the mailing list by Kevin Buckley.  On the rowview we were still setting FFFFFF as a transparent color even though we've redone the model on the racks to not have the department color showing underneath the device anymore and we instead just have the coloring on U designation.  There may be other issues with this that I'm not considering right now but i've removed the conditional check for the row view.  Also it was noted that the draggable positioning wasn't working ont he row view.  This was caused by the code refactoring that was done several months ago to common.js to add in the placeholders among other things.  I added a place for it to call init drag after all the racks have loaded to correct the error