Added in the Spout library, which triggered an update of PHPExcel to the last version before abandonment of the project.
Also cleaned up the mpdf ttfonts to remove the non-free fonts, which
we aren't using in the stock reports.   That dropped size by many MB.

Migrated the Bulk Cabinet importer to the new Spout library.   Easy to
use and is fast (because it has less features), but it reminds me of the
old RPG report writing days.   Access to the Excel data is sequential.

#1105