JS OOO Correction #1111
Found the bug that was causing trouble.  The tabs were getting initialized and hiding the button that the javascript was looking to target.  Moved the bind command inside of hte document ready function but before the tabs are initialized so the object is present to act upon.  Also added a second nested div for the content to be dumped to because the initial div that was targetted was changed to have the titlebar, etc from the dialog function