Function redirect: don't add an URL
If you install the server behind a proxy, you have to use
X_FORWARDED_FOR values to detect the correct hostname.  The detection
also failed when the service was served on a different port than the
standard 443 or 80.
By removing the URL of the server you can prevent this problems.
According to the mozilla documention [1], using the relative URL is
perfectly correct.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Location