Stupid safari
I can't explain it and i've put too much time into it.  Safari is making the table on the devices.php page longer each time the resize function runs is the page isn't wide enough for the content.  This hack will wait for the resize to finish then put the table caption back in the right place.  Any subsequent resizes of the browser window will likely put the caption off in limbo again.  Perhaps a better solution will be to remove the caption aspect completely at some point