Change location of datacenter id determination.  close #1035
When the rowview was original made we didn't have the datacenter ID as part of the cabinet row and we had to determine the correct datacenter based on the id of one of the cabinets.  Since we have the id now might as well use it to eliminate errors from S.U.T. of defining an empty cabinet row and going to it expecting to see something for whatever unknown reason.  Users are dumb.