Update redirect() function
This updates the redirect() function so it hopefully works better when InstallURL is set and openDCIM may have been deployed in a sub-directory in document root. Previously the path() function (now defunct) would figure out the path for the target based on REQUEST_URI but didn't consider that InstallURL may also be set.

For example if openDCIM was installed in a sub-directory of document root called "test" and InstallURL was correctly set to "http://server/test/" (with or without trailing slash), redirect() in combination with path() would redirect the client to "http://server/test/test/".

Now the redirect() function first checks if InstallURL is set. If it is set, this will be the redirect URL with $target appended. If it's not set, it reverts back to using REQUEST_URI to figure out the path as well as determining if the schema is HTTP or HTTPS and using SERVER_NAME to create the redirect URL.

Tested:
- With and without InstallURL set.
- Creating a device.

Not tested:
- Fresh install.
- Upgrade.
- API.