Correction to devices.php for a log viewer button
timlegge asked on irc about a button that used to be on the devices page that would show all the actions for a device.  I swore it never existed but he found a screenshot and a portion of the code.  One of the fixes for a safari rendering bug long ago removed the caption class from the bottom button bar of devices.php and the javascript that injected that function was sending it to a hidden location for any device that wasn't a CDU/PDU.  Changed the target to the bottom button bar to restore the functionality.  Thanks tim