Add new opendcim mail class
I figured the most common usecase for us is that we're setting the same settings repeatedly and just needed an easy way to load them.  So I extended the PHPMailer class and added an init function that will add in all of our default values.  You can still call the setters from the native class to override any values.  As i'm typing this thinking that the way i'm adding in the recipient may need to be rethought just a little bit before all the instances are updated since I think this will cause it to add int he facility manager everytime.  I might need to think about how to extend the send function to eval at that time for the recipient then.