Alter installURL pathing for non-root installs fixes #1245
I haven't found another condition for this bug so this should be a permanent fix.  The issue lies in the baseurl when it isn't root it was adding whatever the path was on top of the baseurl and causing it to double.  so if you installed to test.com/openDCIM/ and set the baseurl to https://test.com/openDCIM/ it would attempt to redirect to test.com/openDCIM/openDCIM/<whatever> in several instances.  We cannot change the baseurl to just the domain because that is used in the email reporting that cannot read a referrer for the base domain or path leaving us with just dealing with the user side of things to mangle the path or put in a different parameter for the base domain and the baseurl and just really make it confusing for the configuration screen.  My solution here is to subject the similar string from the baseinstall url.  the path function kicks back just the path without any executable page in it so we can just subtract that from the baseInstall string and then allow the target function to add it back in without any complicated logic changes.  the only condition this will fail in is if the user has installed to something like openDCIM/other/  and they set the base install to http://test.com/openDCIM/ which would be wrong for the base anyhow.