Change to js for text editor #1431
This appears to be fixed on my test machine in the latest firefox and chrome.  I have not tested edge, opera, or safari.  We should probably seriously discuss what it owuld take to convert all of these notes to markdown, blech, and deal with the other issues these damned notes have caused