add a check serial number button exists
On the devices.php page,
a check serial number button is added. The button is invisible by default and visible when the page is in edit mode (no serial number in the input) or when the page creates a new device.
An AJAX request is sent to check for the existence of a serial number triggered by the serialexist button.