Update devices.php
On the devices.php page,
a serial number verification button has been added. This button is invisible by default and visible when the page is in edit mode (no serial number in the entry when loading) or when creating a new device.
An AJAX request is sent to verify the existence of a serial number, triggered by the serialexist button.