Add mass edit controls for #1622
Added in new mass edit controls for the new port options.  Had to duplicate the data sources into the header because the row wasn't populating the options on page load.  Doing it this way makes sure that if an edit function is called it will update the choices for everything from the api.  If someone is adding in choices it will show up real time for the edits but the page will have to reload for the massedit to show the new options