Checkpoint #1622
All of the controls on the configuration page are in place for creating and updating the new values.  The delete functionality is still in progress, I have the basic modal in place but we need to discuss how to handle some text for that.  Might use another data aspect on the containers like I did the datapath.

the delete does work for items that do not have any associations in the database

will need to create the api routes for all of the other classes just like the powerconnectortypes. I used the paths that were already in the get routes so they need to stay consistent across all the other methods