Checkpoint #1622
This should have all of the functions working now.  I think there is some improvement to be made to the delete functions in the JS but i'll look at that later.  Any of the phrases that need to be translated can be embedded down in the php and keep the js clean using the named data attributes