Removing repeated function call #1622
Since the delete action is needed multiple times making a single function call that works for all of them.  Had to add in some checking to make sure the different values are available to keep it happy.  Everything is set now that we can do all the translation bits and add the rest of the api routes and test